sensor_module/requirements.txt
RPi.GPIO
rpi_ws281x
smbus2
wheel
psycopg2-binary

